Add files via upload
Added protocol loading to the build protocol tab to allow users to specify which thermocycling protocol to use or extraction protocol to uses, etc. Also added the ability to pick the pre-amp thermocycling protocol on the build protocol tab (update to protocol load in the future?). Modified the thermocycle tab, moved the use check boxes and added a settings button to allow fine tuning of the thermocycler trays and clamps.